BD Helpers de tudo, exceto Jogo, Computador, Smartphone. Alterações nas classes e singletons para acomodar os tipos de dados

+++ app/src/main/java/pt/ipleiria/estg/tesp/psi/projsi/sistematrocas/projetosi_android_1718/singletons/SingletonCategorias.java
@@ -49,9 +49,9 @@ public class SingletonCategorias {
         return categorias.contains(updatedCategoria);
     }
 
-    public Categoria searchCategoriaID(Integer id)
+    public Categoria searchCategoriaID(Long id)
     {
-        return categorias.get(id);
+        return categorias.get(id.intValue());
     }
 
 }
