+++ app/src/main/java/pt/ipleiria/estg/tesp/psi/projsi/sistematrocas/projetosi_android_1718/singletons/SingletonTiposRoupa.java
@@ -49,8 +49,8 @@ public class SingletonTiposRoupa {
         return tiposRoupa.contains(updatedTipo);
     }
 
-    public TipoRoupa searchTipoRoupaID(Integer id)
+    public TipoRoupa searchTipoRoupaID(Long id)
     {
-        return tiposRoupa.get(id);
+        return tiposRoupa.get(id.intValue());
     }
 }
