+++ app/src/main/java/pt/ipleiria/estg/tesp/psi/projsi/sistematrocas/projetosi_android_1718/singletons/SingletonUsers.java
@@ -49,8 +49,8 @@ public class SingletonUsers {
         return users.contains(updatedUser);
     }
 
-    public User searchUserID(Integer id)
+    public User searchUserID(Long id)
     {
-        return users.get(id);
+        return users.get(id.intValue());
     }
 }
